fix: use absolute position of headline in flexible layout
Switching between relative and absolute positioning of headline
causes text flickering (infinite render loop where relative position
changes font size that changes insight dimension, causing it change to
absolute, which changes font size that changes insight dimension, and
repeat...) when viewport width is changed to mobile size. It looks
like absolute positioning works in all cases in flexible layouts,
therefore this logic is disabled when new layout engine is enabled.

JIRA: LX-1144
risk: low
