+++ libs/sdk-ui-dashboard/src/presentation/flexibleLayout/DefaultDashboardLayoutRenderer/DashboardLayout.tsx
@@ -124,7 +124,7 @@ export function DashboardLayout<TWidget>(props: IDashboardLayoutRenderProps<TWid
                 />
             );
         },
-        [resizedItemPositions, widgetRenderer],
+        [resizedItemPositions, widgetRenderer, getLayoutDimensions],
     );
 
     const screenSize = useScreenSize();
